// src/assets/Types/Movie.ts
export interface Movie {
    id: number;
    title: string;
    poster_path: string;
    credits: string;
    tagline: string;
    runtime: string;
    cast: any;
    actor_id : number;
    director: string;
    profile_path: string;
    release_date: string;
    overview: string;
    vote_average: number;
}